requirements.txt
# Core Packages
numpy
pandas
scipy
scikit-learn

# Visualization
matplotlib
seaborn


# Jupyter
notebook
ipykernel
ipywidgets

# Utilities
colorama
pytest
black
streamlit

# Machine Learning
# tensorflow-macos==2.10.0; sys_platform == 'darwin' and 'ARM' in platform_version
# tensorflow==2.10.0; sys_platform == 'darwin' and 'ARM' not in platform_version
# tensorflow==2.10.0; sys_platform != 'darwin'
# mlflow==2.1.1
# tpot
# imblearn

# API
# fastapi
# uvicorn
# pytest-asyncio

# Package specific
streamlit-shadcn-ui  
loguru
pdfminer.six

